move dashboard and post pages to main folder
