+++ src/app/(main)/dashboard/page.tsx
@@ -1,10 +1,10 @@
-import { Sidebar } from '../../components/sidebar'
-import { DashboardFilter } from '../../components/DashboardFilter'
-import { DashboardSearch } from '../../components/DashboardSearch'
-import { DashboardTable } from '../../components/DashboardTable'
-import { DashboardProvider } from '../../lib/providers/DashboardProvider'
+import { Sidebar } from '../../../components/sidebar'
+import { DashboardFilter } from '../../../components/DashboardFilter'
+import { DashboardSearch } from '../../../components/DashboardSearch'
+import { DashboardTable } from '../../../components/DashboardTable'
+import { DashboardProvider } from '../../../lib/providers/DashboardProvider'
 import { getStaticPosts } from '~/utils'
-import logo from '../../assets/logo.png'
+import logo from '../../../assets/logo.png'
 import Image from 'next/image'
 import { withPageAuthRequired } from '@auth0/nextjs-auth0'
 import { SIDEBAR_NAVITEMS } from '~/components/sidebar/constants'
