+++ src/app/(main)/post/[slug]/page.tsx
@@ -5,7 +5,7 @@ import { SlugSection } from '~/components/slug-section'
 import { getClient } from '~/lib/sanity.client'
 import { getPost, Post } from '~/lib/sanity.queries'
 
-import logo from '../../../assets/logo.png'
+import logo from '../../../../assets/logo.png'
 import { SIDEBAR_NAVITEMS } from '~/components/sidebar/constants'
 import { Sidebar } from '~/components/sidebar'
